Texture creation and apply material in runtime
Extended road painter wrapper class in order to create textures statically in the content folder so the road painter blueprint can paint the road map into these new textures and save them, instead of having to share the same texture over several maps. The material is now applied on runtime (as soon as the map is launched, either in package mode or play mode in the editor). Will only work if it uses a custom "Road Painter" blueprint, which will be uploaded to the art repository.